chore: 🤖 Move model hook from session to session.index
