+++ addons/rose/app/styles/rose/components/layout/_body-content.scss
@@ -8,7 +8,7 @@
 .rose-layout-body-content {
   display: grid;
   grid-template-areas: 'body body sidebar';
-  grid-template-columns: 1fr 1fr 300px;
+  grid-template-columns: 1fr 1fr fit-content(300px);
   grid-template-rows: auto 1fr;
   gap: 2rem;
