+++ ui/desktop/app/templates/scopes/scope/projects/sessions/index.hbs
@@ -18,7 +18,7 @@
       <toolbar.primary>
         <Rose::Form as |form|>
           {{#let
-            (resource-filter 'scopes.scope.projects.sessions' 'status')
+            (resource-filter 'scopes.scope.projects.sessions.index' 'status')
             as |filter|
           }}
             <Rose::Dropdown
@@ -45,7 +45,7 @@
           {{/let}}
 
           {{#let
-            (resource-filter 'scopes.scope.projects.sessions' 'project')
+            (resource-filter 'scopes.scope.projects.sessions.index' 'project')
             as |filter|
           }}
             <Rose::Dropdown
@@ -95,7 +95,7 @@
 
           {{#if
             (has-resource-filter-selections
-              'scopes.scope.projects.sessions' 'status' 'project'
+              'scopes.scope.projects.sessions.index' 'status' 'project'
             )
           }}
             <Rose::Button
@@ -143,7 +143,7 @@
                   @acknowledgeText={{t 'states.copied'}}
                 >
                   <LinkTo
-                    @route='scopes.scope.projects.sessions.session'
+                    @route='scopes.scope.projects.sessions.session.index'
                     @model={{B.data.id}}
                     data-test-session-detail-link
                   >
@@ -215,7 +215,7 @@
       </Hds::Table>
 
     {{else if (has-resource-filter-selections
-      'scopes.scope.projects.sessions' 'status' 'project'
+      'scopes.scope.projects.sessions.index' 'status' 'project'
     )}}
       <Rose::Layout::Centered>
         <Rose::Message @title={{t 'titles.no-results'}} as |message|>
